finish rewrite
