+++ pkg/broker/consumer_group.go
@@ -2,19 +2,17 @@ package broker
 
 import (
 	"fmt"
-	"log"
-	"sort"
 	"sync"
 	"time"
 
+	"github.com/alexandrecolauto/gofka/pkg/topic"
 	"github.com/alexandrecolauto/gofka/proto/broker"
 )
 
 type ConsumerGroup struct {
 	id       string
 	leaderId string
-	topics   map[string]*Topic
-	offsets  map[OffsetKey]int
+	topics   map[string]*topic.Topic
 
 	topicList  map[string]bool
 	consumers  map[string]*ConsumerSession
@@ -25,20 +23,13 @@ type ConsumerGroup struct {
 	mu         sync.RWMutex
 }
 
-type OffsetKey struct {
-	GroupID   string
-	Topic     string
-	Partition int
-}
-
 func NewConsumerGroup(id string) *ConsumerGroup {
 	co := make(map[string]*ConsumerSession)
-	ts := make(map[string]*Topic)
+	ts := make(map[string]*topic.Topic)
 	tl := make(map[string]bool)
-	of := make(map[OffsetKey]int)
 	c_ch := make(chan *ConsumerSession)
 	d_ch := make([]chan any, 0)
-	return &ConsumerGroup{id: id, topics: ts, consumers: co, offsets: of, consumerCh: c_ch, doneChList: d_ch, topicList: tl}
+	return &ConsumerGroup{id: id, topics: ts, consumers: co, consumerCh: c_ch, doneChList: d_ch, topicList: tl}
 }
 
 func (cg *ConsumerGroup) ResetConsumerGroup(doneCh chan any) {
@@ -46,7 +37,6 @@ func (cg *ConsumerGroup) ResetConsumerGroup(doneCh chan any) {
 	if cg.joining {
 		return
 	}
-	fmt.Println("Consumer group reseted")
 	co := make(map[string]*ConsumerSession)
 	cg.consumers = co
 	cg.joining = true
@@ -55,7 +45,6 @@ func (cg *ConsumerGroup) ResetConsumerGroup(doneCh chan any) {
 	for {
 		select {
 		case con := <-cg.consumerCh:
-			fmt.Println("New consumer joined: ", con)
 			if cg.leaderId == "" {
 				cg.leaderId = con.id
 			}
@@ -65,18 +54,12 @@ func (cg *ConsumerGroup) ResetConsumerGroup(doneCh chan any) {
 
 		case <-timeout.C:
 			for _, d_ch := range cg.doneChList {
-				fmt.Println("finished ")
 				close(d_ch)
 			}
 			cg.joining = false
+			cg.doneChList = make([]chan any, 0)
 			return
 		}
-	}
-}
-
-func (cg *ConsumerGroup) AddTopics(topics []string) {
-	for _, tp := range topics {
-		cg.topicList[tp] = true
 	}
 }
 
@@ -92,110 +75,46 @@ func (cg *ConsumerGroup) AddConsumer(id string, topics []string) {
 	cg.consumerCh <- c
 }
 
-func (cg *ConsumerGroup) GetRegisterResponse() *broker.RegisterConsumerResponse {
+func (cg *ConsumerGroup) GetRegisterResponse(id string) *broker.RegisterConsumerResponse {
+	cg.mu.RLock()
+	defer cg.mu.RUnlock()
 	topicList := make([]string, 0)
 	for topic := range cg.topicList {
 		topicList = append(topicList, topic)
 	}
-	conumers := make([]*broker.ConsumerSession, 0)
-	for _, consumer := range cg.consumers {
-		conumers = append(conumers, &broker.ConsumerSession{
-			Id:     consumer.id,
-			Topics: consumer.topics,
-		})
+	conumers_list := []*broker.ConsumerSession{}
+	for c_id := range cg.consumers {
+		c := cg.consumers[c_id]
+		newConsumer := &broker.ConsumerSession{
+			Id:     c.id,
+			Topics: c.topics,
+		}
+		conumers_list = append(conumers_list, newConsumer)
 	}
 	res := &broker.RegisterConsumerResponse{
 		Success:   true,
 		Leader:    cg.leaderId,
 		AllTopics: topicList,
-		Consumers: conumers,
+		Consumers: conumers_list,
 	}
 	return res
 }
 
-func (cg *ConsumerGroup) RemoveConsumer(id string, topic *Topic) {
-	cg.mu.Lock()
-	defer cg.mu.Unlock()
-
-	if _, ok := cg.consumers[id]; ok {
-		delete(cg.consumers, id)
-		if len(cg.consumers) > 0 {
-			cg.rebalance()
-		}
+func (cg *ConsumerGroup) AddTopics(topics []string) {
+	for _, tp := range topics {
+		cg.topicList[tp] = true
 	}
 }
 
-func (cg *ConsumerGroup) rebalance() {
-	if len(cg.consumers) == 0 {
+func (cg *ConsumerGroup) Subscribe(topic *topic.Topic) {
+	_, ok := cg.topics[topic.Name]
+	if ok {
 		return
 	}
 
-	var all_partitions []*broker.PartitionInfo
-	for topicName, topic := range cg.topics {
-		for i := range topic.n_partitions {
-			all_partitions = append(all_partitions, &broker.PartitionInfo{Id: int32(i), TopicName: topicName})
-
-		}
-	}
-
-	sort.Slice(all_partitions, func(i, j int) bool {
-		if all_partitions[i].TopicName == all_partitions[j].TopicName {
-			return all_partitions[i].Id < all_partitions[j].Id
-		}
-		return all_partitions[i].TopicName < all_partitions[j].TopicName
-	})
-
-	consumerIDs := make([]string, 0, len(cg.consumers))
-	for id := range cg.consumers {
-		consumerIDs = append(consumerIDs, id)
-	}
-	sort.Strings(consumerIDs)
-
-	for _, consumer := range cg.consumers {
-		consumer.partitions = make([]*broker.PartitionInfo, 0)
-	}
-
-	for i, partition := range all_partitions {
-		consumerIdx := i % len(consumerIDs)
-		consumerID := consumerIDs[consumerIdx]
-		consumer := cg.consumers[consumerID]
-		consumer.partitions = append(consumer.partitions, partition)
-		fmt.Printf("Assigning partition %v to user\n", partition)
-	}
-}
-
-func (cg *ConsumerGroup) GetAssignedPartitions(consumerID string) []*broker.PartitionInfo {
-	cg.mu.RLock()
-	defer cg.mu.RUnlock()
-
-	if consumer, ok := cg.consumers[consumerID]; ok {
-		partitions := make([]*broker.PartitionInfo, len(consumer.partitions))
-		copy(partitions, consumer.partitions)
-		return partitions
-	}
-	return nil
-}
-
-func (cg *ConsumerGroup) FetchMessages(id string, opt *broker.ReadOptions) ([]*broker.Message, error) {
-	c := cg.consumers[id]
-	var msgs []*broker.Message
-	for _, part := range c.partitions {
-		topic := cg.topics[part.TopicName]
-		log.Println("fetching msg for:", part.TopicName)
-		key := OffsetKey{Topic: part.TopicName, Partition: int(part.Id), GroupID: cg.id}
-		offset := cg.offsets[key]
-		var items []*broker.Message
-		items, err := topic.ReadFromPartition(int(part.Id), offset, opt)
-		if err != nil {
-			return nil, err
-		}
-		msgs = append(msgs, items...)
-	}
-	fmt.Println("Final messages", len(msgs), len(c.partitions))
-	if len(msgs) > 0 {
-		return msgs, nil
-	}
-	return nil, nil
+	cg.mu.Lock()
+	defer cg.mu.Unlock()
+	cg.topics[topic.Name] = topic
 }
 
 func (cg *ConsumerGroup) SyncGroup(consumers []*broker.ConsumerSession) {
@@ -206,6 +125,7 @@ func (cg *ConsumerGroup) SyncGroup(consumers []*broker.ConsumerSession) {
 	}
 	cg.inSync = true
 }
+
 func (cg *ConsumerGroup) UserAssignment(user_id string, retries int) (*broker.ConsumerSession, error) {
 	if !cg.inSync {
 		time.Sleep(time.Duration(retries) * 100 * time.Millisecond)
@@ -231,31 +151,6 @@ func (cg *ConsumerGroup) UserAssignment(user_id string, retries int) (*broker.Co
 	return cs, nil
 }
 
-func (cg *ConsumerGroup) UpdateOffset(topic string, partition, offset int) {
-	key := OffsetKey{
-		Topic:     topic,
-		Partition: partition,
-		GroupID:   cg.id,
-	}
-	fmt.Printf("New offset for %s, p: %d, offset: %d\n", topic, partition, offset)
-	cg.offsets[key] = offset
-}
-
-func (cg *ConsumerGroup) Subscribe(topic *Topic) {
-	cg.mu.Lock()
-	defer cg.mu.Unlock()
-
-	cg.topics[topic.name] = topic
-	cg.rebalance()
-}
-func (cg *ConsumerGroup) Unsubscribe(topic_name string) {
-	cg.mu.Lock()
-	defer cg.mu.Unlock()
-
-	delete(cg.topics, topic_name)
-	cg.rebalance()
-}
-
 func (cg *ConsumerGroup) ConsumerHeartbeat(id string) {
 	cg.mu.Lock()
 	defer cg.mu.Unlock()
@@ -264,31 +159,18 @@ func (cg *ConsumerGroup) ConsumerHeartbeat(id string) {
 		c.last_heartbeat = time.Now()
 	}
 }
-
-func (cg *ConsumerGroup) UnregisterConsumer(id string) {
-	cg.mu.Lock()
-	defer cg.mu.Unlock()
-	delete(cg.consumers, id)
-	cg.rebalance()
-}
-
 func (cg *ConsumerGroup) ClearDeadSession() {
 	cg.mu.Lock()
 	defer cg.mu.Unlock()
 	deadSessionTimeout := 15 * time.Second
-	now := time.Now()
 	for _, consumer := range cg.consumers {
-		if now.Sub(consumer.last_heartbeat) > deadSessionTimeout {
+		if time.Since(consumer.last_heartbeat) > deadSessionTimeout {
 			cg.UnregisterConsumer(consumer.id)
 		}
 	}
 }
-
-func (cg *ConsumerGroup) DeleteTopic(topic_name string) {
-	if _, ok := cg.topics[topic_name]; !ok {
-		fmt.Println("cant find topic")
-		return
-	}
-	delete(cg.topics, topic_name)
-	cg.rebalance()
+func (cg *ConsumerGroup) UnregisterConsumer(id string) {
+	cg.mu.Lock()
+	defer cg.mu.Unlock()
+	delete(cg.consumers, id)
 }
